Multiple changes rename of project
Changed to name of the program to fit with our selected icon. Yeah that is sort of lame, but it still fits with the name goal which is that I don’t want the name to sound like the trillion other password generators which are named:
passgen, pass keeper, passwordvalut, secretvalt boa bla. Very predictable names and hard to tell apart. Everyone has icons of vaults, keys or keyholes.