Started using bindings for length and series
I wanted to avoid this initially, but bindings simplify code a lot when you need to keep two fields in synch. This was less of a problem when I only had a label and not a textfield for length and series.
setNilValueForKey: was important to get right otherwise removing all numbers in textfield would cause exception.