Got unit tests to compile rething TLLogin
We probably don’t want TLPasswordGenerator to manage TLLogin arrays. That makes it unclear what it actually does. Secondly it is better that generating a password creates a TLLogin object, then all the relevant data which belong together is put together. This reduces chance of screwing up and bundling passwords with the wrong username and target.