Make it possible to remember previous password length
This made us revert the use of bindings again. The problem was that when binding to NSUserDefaults keys we couldn't figure out how to make the bindings understand we are dealing with ints and not floats. So the slider would produce float values.

The result is code that is a lot more verbose. All of this will need to be cleaned up at some point.