Added toolbar support so we can have a place for accessing more auxiliary functions.
Motivation: I want to provide a way for user to view stored logins and validate them. I don’t want to mess up the main window with buttons for that. Thus I am adding a toolbar. I am disabling the button for showing logins until I have a window for it.