Fiddling with getting QR code generation working
An exception is thrown when hitting the print button.

My hunch is that having a Window Controller load and manage the view we want to print might not be a good idea. When I just created a view from scratch I didn't get this problem. But I'd like to use Inteface builder to handle layout. The MIHQRCodeView has been changed too much to call it an external lib anymore. Should probably rename and maintain that as my own class.