Printing seems to work.
This won't actually catch the __cxa_throw exception
But we don't need to. "__cxa_throw" is from the C++ subsystem and was likely and exception which the subsystem caught and handled. Just part of normal error handling. However if we set breakpoint on all exceptions, we catch all exceptions. It doesn't matter if the exception is handled or not. I is easy to think that the breakpoint is just there to catch uncaught exceptions. That is not the case.